pyproject.toml
[tool.poetry]
name = "pythonproject1"
version = "0.1.0"
description = ""
authors = ["ryuga <vaishnav2353@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
starlette = "^0.17.1"
tortoise-orm = "^0.17.8"
uvicorn = "^0.16.0"
asyncpg = "^0.25.0"
aioredis = "^2.0.0"
httpx = "0.16.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
